requirements.txt
anyio==3.6.2
Authlib==1.2.0
bidict==0.22.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
dnspython==2.3.0
fastapi==0.92.0
fastapi-socketio==0.0.10
fonttools==4.39.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.21.0
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
markdown2==2.4.8
MarkupSafe==2.1.2
matplotlib==3.7.1
nicegui==1.1.10
numpy==1.24.2
orjson==3.8.7
packaging==23.0
Pillow==9.4.0
plotly==5.13.1
protobuf==4.22.1
pscript==0.7.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.6
Pygments==2.14.0
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.3.4
python-multipart==0.0.5
python-socketio==5.7.2
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tenacity==8.2.2
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.20.0
vbuild==0.8.1
watchfiles==0.18.1
websockets==10.4
